rename appbar files -> NavBar
appBar.css -> navBar.css
AppBar.tsx -> NavBar.tsx

updated layout.tsx to import from newly named
 NavBar file
